TASK: Extract the cache flushing from the parser cache
That way not the parser cache is connected in the packahe.php.
The consistency of the cache identifiers is ensured by using a ParserCacheIdentifierTrait that is used by the parserCache and the ParserCacheFlusher.
