+++ Neos.Fusion/Classes/Core/Cache/ParserCache.php
@@ -17,8 +17,6 @@ use Neos\Flow\Annotations as Flow;
 use Neos\Cache\Frontend\VariableFrontend;
 use Neos\Flow\Package\PackageManager;
 use Neos\Fusion\Core\ObjectTreeParser\Ast\FusionFile;
-use Neos\Utility\Unicode\Functions as UnicodeFunctions;
-use Neos\Utility\Files;
 
 /**
  * Helper around the ParsePartials Cache.
@@ -28,6 +26,8 @@ use Neos\Utility\Files;
  */
 class ParserCache
 {
+    use ParserCacheIdentifierTrait;
+
     /**
      * @Flow\Inject
      * @var VariableFrontend
@@ -66,19 +66,6 @@ class ParserCache
         return $this->cacheForIdentifier($identifier, $generateValueToCache);
     }
 
-    /**
-     * @param array<string, int> $changedFiles
-     */
-    public function flushFileAstCacheOnFileChanges(array $changedFiles): void
-    {
-        foreach ($changedFiles as $changedFile => $status) {
-            $identifier = $this->getCacheIdentifierForFile($changedFile);
-            if ($this->parsePartialsCache->has($identifier)) {
-                $this->parsePartialsCache->remove($identifier);
-            }
-        }
-    }
-
     private function cacheForIdentifier(string $identifier, \Closure $generateValueToCache): mixed
     {
         if ($this->parsePartialsCache->has($identifier)) {
@@ -87,58 +74,5 @@ class ParserCache
         $value = $generateValueToCache();
         $this->parsePartialsCache->set($identifier, $value);
         return $value;
-    }
-
-    /**
-     * creates a comparable hash of the dsl type and content to be used as cache identifier
-     */
-    private function getCacheIdentifierForDslCode(string $identifier, string $code): string
-    {
-        return 'dsl_' . $identifier . '_' . md5($code);
-    }
-
-    /**
-     * creates a comparable hash of the absolute, resolved $fusionFileName
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function getCacheIdentifierForFile(string $fusionFileName): string
-    {
-        if (str_contains($fusionFileName, '://')) {
-            $fusionFileName = $this->getAbsolutePathForPackageRessourceUri($fusionFileName);
-        }
-
-        $realPath = realpath($fusionFileName);
-        if ($realPath === false) {
-            throw new \InvalidArgumentException("Couldn't resolve realpath for: '$fusionFileName'");
-        }
-
-        $realFusionFilePathWithoutRoot = str_replace(FLOW_PATH_ROOT, '', $realPath);
-        return 'file_' . md5($realFusionFilePathWithoutRoot);
-    }
-
-    /**
-     * Uses the same technique to resolve a package resource URI like Flow.
-     *
-     * resource://My.Site/Private/Fusion/Foo/Bar.fusion
-     * ->
-     * FLOW_PATH_ROOT/Packages/Sites/My.Package/Resources/Private/Fusion/Foo/Bar.fusion
-     *
-     * {@see \Neos\Flow\ResourceManagement\Streams\ResourceStreamWrapper::evaluateResourcePath()}
-     * {@link https://github.com/neos/flow-development-collection/issues/2687}
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function getAbsolutePathForPackageRessourceUri(string $requestedPath): string
-    {
-        $resourceUriParts = UnicodeFunctions::parse_url($requestedPath);
-
-        if ((isset($resourceUriParts['scheme']) === false
-            || $resourceUriParts['scheme'] !== 'resource')) {
-            throw new \InvalidArgumentException("Unsupported stream wrapper: '$requestedPath'");
-        }
-
-        $package = $this->packageManager->getPackage($resourceUriParts['host']);
-        return Files::concatenatePaths([$package->getResourcesPath(), $resourceUriParts['path']]);
     }
 }
