+++ Neos.Fusion/Configuration/Objects.yaml
@@ -27,3 +27,13 @@ Neos\Fusion\Core\Cache\ParserCache:
         arguments:
           1:
             value: Neos_Fusion_ParsePartials
+
+Neos\Fusion\Core\Cache\ParserCacheFlusher:
+  properties:
+    parsePartialsCache:
+      object:
+        factoryObjectName: Neos\Flow\Cache\CacheManager
+        factoryMethodName: getCache
+        arguments:
+          1:
+            value: Neos_Fusion_ParsePartials
